Moving database access functions out of login
